api-credito/requirements.txt
# FastAPI stack
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# HTTP client (para integração com banco mock)
httpx==0.27.0

# AWS SQS
boto3==1.34.113

# Async utilities
tenacity==8.3.0

# Validation
email-validator==2.1.1

# Environment variables
python-dotenv==1.0.1

# Logging
structlog==24.1.0

# Testing
pytest==8.2.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0

# Monitoring
prometheus-fastapi-instrumentator

